requirements.txt
Crawl4AI==0.4.247
fake-http-header==0.3.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
googleapis-common-protos==1.66.0
narwhals==1.21.1
nltk==3.9.1
numpy==2.2.1
openai==1.59.6
packaging==24.2
pandas==2.2.3
pillow==10.4.0
playwright==1.49.1
pluggy==1.5.0
protobuf==5.29.3
psutil==6.1.1
pyarrow==18.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
realtime==2.1.0
referencing==0.35.1
rsa==4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
watchdog==6.0.0
websockets==13.1
zipp==3.21.0